Use the same CallIdSequence on both client and member sides
